fix(components): merge vue virtual-list
