+++ packages/taro-components-advanced/src/components/virtual-list/index.ts
@@ -6,6 +6,19 @@ interface VirtualListProps extends Omit<StandardProps, 'children'> {
   height: string | number
   /** 列表的宽度。 */
   width: string | number
+  /** 子组件 */
+  item: ComponentType<{
+    /** 组件 ID */
+    id: string
+    /** 单项的样式，样式必须传入组件的 style 中 */
+    style?: CSSProperties
+    /** 组件渲染的数据 */
+    data: any
+    /** 组件渲染数据的索引 */
+    index: number
+    /** 组件是否正在滚动，当 useIsScrolling 值为 true 时返回布尔值，否则返回 undefined */
+    isScrolling?: boolean
+  }>
   /** 列表的长度 */
   itemCount: number
   /** 渲染数据 */
@@ -49,21 +62,9 @@ interface VirtualListProps extends Omit<StandardProps, 'children'> {
   overscanCount?: number
   /** 上下滚动预占位节点 */
   placeholderCount?: number
-  /** 是否注入 isScrolling 属性到 children 组件。这个参数一般用于实现滚动骨架屏（或其它 placeholder） 时比较有用。 */
+  /** 是否注入 isScrolling 属性到 item 组件。这个参数一般用于实现滚动骨架屏（或其它 placeholder） 时比较有用。 */
   useIsScrolling?: boolean
   style?: CSSProperties
-  children?: ComponentType<{
-    /** 组件 ID */
-    id: string
-    /** 单项的样式，样式必须传入组件的 style 中 */
-    style?: CSSProperties
-    /** 组件渲染的数据 */
-    data: any
-    /** 组件渲染数据的索引 */
-    index: number
-    /** 组件是否正在滚动，当 useIsScrolling 值为 true 时返回布尔值，否则返回 undefined */
-    isScrolling?: boolean
-  }>
 }
 
 declare namespace VirtualListProps {
@@ -117,12 +118,11 @@ declare namespace VirtualListProps {
  *       <VirtualList
  *         height={500} // 列表的高度
  *         width='100%' // 列表的宽度
+ *         item={Row} // 列表单项组件，这里只能传入一个组件
  *         itemData={data} // 渲染列表的数据
  *         itemCount={dataLen} // 渲染列表的长度
  *         itemSize={100} // 列表单项的高度
- *       >
- *         {Row} // 列表单项组件，这里只能传入一个组件
- *       </VirtualList>
+ *       />
  *     );
  *   }
  * }
@@ -187,17 +187,8 @@ declare namespace VirtualListProps {
  */
 declare class VirtualListComponent extends Component<VirtualListProps> {}
 
-function getVirtualListComponent () {
-  switch (process.env.FRAMEWORK) {
-    case 'vue3':
-      return require('./vue/vue3').default
-    case 'vue':
-      return require('./vue').default
-    default:
-      return require('./react').default
-  }
-}
-const VirtualList: typeof VirtualListComponent = getVirtualListComponent()
+const VirtualList: typeof VirtualListComponent = require('./react').default
+export const registerVirtualList = require('./vue').default
 
 export { VirtualList, VirtualListProps }
 export default VirtualList
