+++ packages/taro-components/virtual-list/index.js
@@ -1,3 +1,4 @@
-module.exports = require('@tarojs/components-advanced/dist/components/virtual-list/index.js').default
+const VirtualList = require('@tarojs/components-advanced/dist/components/virtual-list/index.js')
 
-module.exports.default = module.exports
+module.exports = VirtualList
+module.exports.default = VirtualList.default
